fixed the readme, script, file menegment
